requirements.txt
Django~=4.2.13
pytz~=2024.1
djangorestframework~=3.15.2
pymongo~=4.8.0
python-dotenv~=1.0.1